requirements.txt
PySide6>=6.6.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
